+++ index.css
@@ -48,3 +48,7 @@
     height: 130px;
     
 }
+.filter-btn {
+    margin: 10px;
+    cursor: pointer;
+}
