+++ index.html
@@ -12,7 +12,7 @@
         <h1>Not Rendered!</h1>
     </div>
     
-    <script type = "module" src="./App.js">
+    <script type = "module" src="./src/App.js">
        
     </script>
 </body>
