+++ src/App.js
@@ -1,0 +1,36 @@
+ import React from "react";
+ import ReactDOM from "react-dom/client";
+ import Header from "./components/Header";
+ import Body from "./components/Body";
+
+
+
+// styles
+const StyleCard = () =>{
+    backgroundcolor :"black"
+}
+
+
+const AppLayout = () => {
+
+    return (
+        <div className="app">
+            <Header/>
+            <Body/>
+        </div>
+    )
+}
+
+
+
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+
+
+root.render(<AppLayout/>);
+
+// JSX
+
+
+
