+++ src/components/Header.js
@@ -1,0 +1,19 @@
+const Header = () => {
+    return (
+        <div className="header">
+            <div className="logo-container">
+                <img  className="logo"  img="/Assets/BiharEssence.jpeg" alt="Logo"/>
+            </div>
+            <div className="nav-items">
+                <ul>
+                    <li>Home</li>
+                    <li>About Us</li>
+                    <li>Contact Us</li>
+                    <li>Cart</li>
+                </ul>
+            </div>
+
+        </div>
+    )
+};
+export default Header;  // standard practice
