+++ utils/constants.js
@@ -1,0 +1,5 @@
+
+
+export const CDN_URL = "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/" ;
+export const LOGO_URL = "" ;
+
